+++ bloom_nofos/users/managers.py
@@ -4,8 +4,8 @@ from django.utils.translation import gettext_lazy as _
 
 class BloomUserManager(BaseUserManager):
     """
-    Custom user model manager where email is the unique identifiers
-    for authentication instead of usernames.
+    Custom user model manager where email is the unique identifier
+    for authentication instead of username.
     """
 
     def create_user(self, email, password, **extra_fields):
@@ -15,6 +15,14 @@ class BloomUserManager(BaseUserManager):
         if not email:
             raise ValueError(_("Email is required"))
         email = self.normalize_email(email)
+
+        group = extra_fields.get("group", "")
+        if not group:
+            raise ValueError(_("Group is required"))
+
+        if group != "bloom" and extra_fields.get("is_staff"):
+            raise ValueError(_("Non-bloom users cannot be staff users."))
+
         user = self.model(email=email, **extra_fields)
         user.set_password(password)
         user.save()
@@ -27,7 +35,10 @@ class BloomUserManager(BaseUserManager):
         extra_fields.setdefault("is_staff", True)
         extra_fields.setdefault("is_superuser", True)
         extra_fields.setdefault("is_active", True)
+        extra_fields.setdefault("group", "bloom")
 
+        if extra_fields.get("group") != "bloom":
+            raise ValueError(_("Superuser must have group=Bloomworks."))
         if extra_fields.get("is_staff") is not True:
             raise ValueError(_("Superuser must have is_staff=True."))
         if extra_fields.get("is_superuser") is not True: