+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 
 ### Added
 
+- Add "group" to NOFOs (represents the OpDiv the NOFO is associated with)
+- Add "group" to users (represents the OpDiv that the user comes from)
+
 ### Changed
 
 ### Fixed